alofans-back/pyproject.toml
[tool.poetry]
name = "AloFa"
version = "0.1.0"
description = "Projeto AlõFã"
authors = ["Mauricio Benjamin da Rocha", "Junior Prado", "Alexandre Lira"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic = "^2.6.1"
passlib = "^1.7.4"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pg8000 = "^1.30.4"
sqlalchemy = "^2.0.27"
python-jose = "^3.3.0"
uuid = "^1.30"
httpx = "^0.27.0"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
pillow = "^10.4.0"
sse-starlette = "^2.1.3"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]
pythonpath = [
  "."
]